feat(tgi server): make the TGI server work without neuron
The code the server is based on is based on the optimum neuron code, in
particular it uses many features of the NeuronModelForCausalLM.
This commit should allow to use the server on any AutoModelForCausalLM,
without the neuron requirements.
Notable changes are the shallow modelling implementation (just a
wrapper), KV cache and positional ids handling, tests adaptations.

This makes most of the server tests pass correctly, not the integration
tests though.
On some platforms, all tests with "sample" do not work as expected for
now.