feat(xla): enable compilation on XLA for TPU support for tgi-server
TGI server now compiles models and runs them on the XLA backend.
The compilation is quite slow and it might not be perfect, but it is a
first step that starts supporting TPUs.

Note that when doing this, a test now fails, probably due to an issue
with XLA compilation, to be fixed later.