Basic TGI server on XLA (#1)
* chore: import gitignore from optimum-neuron

* chore: import text-generation-inference from optimum-neuron

The code will actually going to be modified later, but this commit will
allow to better see differences from the original code.

Reference version of the original code is v0.0.18.

* chore(tgi server): import logits_process and token_selector from neuron

More files copied over, these will be used too.

* doc: add basic README

* feat(tgi server): make the TGI server work without neuron

The code the server is based on is based on the optimum neuron code, in
particular it uses many features of the NeuronModelForCausalLM.
This commit should allow to use the server on any AutoModelForCausalLM,
without the neuron requirements.
Notable changes are the shallow modelling implementation (just a
wrapper), KV cache and positional ids handling, tests adaptations.

This makes most of the server tests pass correctly, not the integration
tests though.
On some platforms, all tests with "sample" do not work as expected for
now.

* fix(pyproject): correct dependencies, description and author

* feat: docker image for TPU TGI server now works

It is now possible to build the image by calling `make tpu-tgi`

* chore: rename neuron -> tpu

To avoid confusion in the future. The only part left untouched is the
integration tests directory, because these tests have not been
adapted yet.

* feat: bump dependency versions (tgi 1.4.2, protobuf)

This also fixes build on macOS.

* test: remove do_sample to increase reproducibility

It seems that different systems use different generators. This leads to
different results and failing tests when using do_sample (even if logits
are close).
Tests using do_sample are removed, replaced by other variations of
parameters.

* feat(xla): enable compilation on XLA for TPU support for tgi-server

TGI server now compiles models and runs them on the XLA backend.
The compilation is quite slow and it might not be perfect, but it is a
first step that starts supporting TPUs.

Note that when doing this, a test now fails, probably due to an issue
with XLA compilation, to be fixed later.

* chore(docker): update rust version

* chore(docker): add _VERSION suffixes for clarity

* chore: simplify model fetching and instantiation

No need to snapshot_download anymore.

* chore: rename modelling -> modeling

To be more coherent with other 🤗 projects.

* Revert "test: remove do_sample to increase reproducibility"

This reverts commit 2eb0958db7465d19b505d02fd9a2b368bbfe6186.

It actually seems like a good idea to have tests with do_samples.
A commit will follow that will adapt tests to the platform where they
are supposed to run.

* fix(tests): Tests are now adapted to TPU/XLA results

* chore: update transformers minor version

* chore: rephrase README

* chore: weights download and model instantiation separated again

* fix(generator): correct generator max_length -> n_positions

* chore(Dockerfile): add missing python dependency

* chore(tests): update gpt2 model id in tests

* feat(tests): docker integration tests adapted to TPU

* fix(test): remove superflous print statements