feat: use dynamic batching when generating
This will increment the number of batches dynamically when you call
prefill, and it will reduce the number of batches only when prefill is
called again.
The intention is to avoid useless recompilation (keeping batch size the
same as long as possible). Also note that if a slot was removed, the
whole KV cache should be rebuilt, and for now we do not do that.