feat(generator): add sharding support for models
All models are now loaded in a sharded way, with each shard being loaded
in a separate process that will run in parallel! 🔀

This allows for better memory management, in particular for larger
models that would not fit in one TPU.

`TpuGenerator` class is now renamed `TpuGeneratorSingleThread`, and a
new `TpuGenerator` class is introduced that will handle the processes
communication.