Sharding in tgi (#31)
* feat: reduce debug messages in distributed model

* fix(style): correct ruff lint entries in pyproject and fix style

* chore: bump version to 0.1.0a0

Versions devX with X > 0 are not supported. Also, it would be better to
synchronize optimum tpu and TGI versions.

* chore(tgi): refactor generator to separate base class

* refactor: split distributed model and mailboxes

This will allow the reuse of the mailboxes in other parts of the code.

* chore(distributed): remove model config from Distributed model

It's not necessary and removing it makes the mailboxes more generic.

* refactor(multiprocessing): allow flexible parameters

- Allow command_data to contain an arbitrary amount of parameters. This
  will allow more flexibility in upcoming uses.
- remove "model" mention from the mailboxes variables, to allow them to
  keep being as generic as possible.

* feat(generator): add sharding support for models

All models are now loaded in a sharded way, with each shard being loaded
in a separate process that will run in parallel! 🔀

This allows for better memory management, in particular for larger
models that would not fit in one TPU.

`TpuGenerator` class is now renamed `TpuGeneratorSingleThread`, and a
new `TpuGenerator` class is introduced that will handle the processes
communication.

* feat: reduce log messages in multiprocess generator

Only ordinal with rank 0 will now log messages.

* test(tgi): add a test for gemma-7b in TGI

Note that this test will only run with --runslow parameter set.

* chore(MP): Add documentation and return type hints for mailboxes

* review(distributed_model): avoid checking len