Basic Llama2 Tuning (#39)
* chore: import run_clm from transformers

This will be then tweaked for TPU/XLA.
Original transformers version is 4.40.0, commit 745bbfe.

* feat(llama): torch_dtype is now bfloat16

This should reduce memory consumption, with low performance loss.

* chore(llama): added changes to prepare for spmd

* chore: import trainer from transformers

Imported as-is, from version 4.39.3.

* feat(xla model parallel): use Linear on SPMD

For RowParallelLinear and ColumnParallelLinear use Linear instead of the
dedicated class, to avoid issues with the backward step.

* feat(llama): pick Linear implementation using `create`

* fix(llama): disable cache when using SPMD

For now it does not seem to work, seems to be related to shapes of key
states that are not compatible with the attention calculation after
update. We can investigate the reason and propose a better solution
later.

* feat(training): adapt trainer and run_clm to optimum and SPMD

* feat(examples): add a short example on llama fine-tuning

* feat(examples): add Llama3-8b tuning instructions

This part uses the multihost environment, tested on v5e litepod16.