Update to Pytorch 2.3.0 and transformers v4.40.2 (#41)
* chore: update pytorch/pytorch xla to 2.3.0

* chore: update transformers version to 4.40.2

* fix(generator): add a dimension to stopping criteria

Now EosTokenCriteria in transformers requires inputs to have at least 2
dimensions, but the TGI generator was giving only one. Unsqueezing those
seem to fix the issue.

* feat(ci): move docker images to pt2.3.0

* chore(ci): update docker image tag again

Setting it to current nightly, because earlier versions do not see the
libtpu and fail.
Also the basic python test that checks if XLA device is removed, since
it was flawed beccause it could fall back to CPU and not detecting
anything, and also because the `make tgi_test` command in the workflow
was re-installing a different torch xla and libtpu versions anyway.