Fine tuning with FSDP v2 (#44)
* feat(training): tuning is now done by using FSDP v2

Custom SPMD sharding is more complicated and does not offer signiificant
performance advantages over FSDP v2 (that uses SPMD).
Example is updated and two helpers have been added to optimum.tpu.

* chore(training): remove unused specific trainer class

SPMD sharding is not used anymore in training, unless FSDP v2 is used,
that is supported by transformers and pytorch 2.3. So this is now
deleted.

* chore(llama): revert changes to prepare for spmd

No special need in modeling, since there is global mesh and shard
marking is done in the Trainer class now.

* feat(gemma): use Linear when in SPMD/FSPD v2

* feat(fsdp v2): training arguments are deduced from model

Trying to make optimum tpu smart enough to deduce fine tune args for
supported models.

* feat(training): added Gemma training example notebook

* doc(readme): update the training section

* feat(tuning): gemma tuning works well on gemma 2b

For gemma 7b a bigger system might be necessary to avoid OOMs.

* fix(train): fsdpv2 training arguments deduced also for PEFT models

Also if model is not matched, raise a clear error.