fix(tgi): handle invalid generation config error and return to server
If there is an invalid generation config, the selector raises an error.
This is caught by the prefill method, that skips the slot generation,
so the error is handled by the router.
I had not been able to reproduce the problem with a simple HTTP request
to TGI, but it seems it's possible to do it with the HTML form
interface, so it's better to handle this, even if it's unlikely to
happen.