Correct extra token, start preparing docker image for TGI/Jetstream Pt (#93)
* fix(Jetstream Pt): remove extra token in decode

Jetstream's `generate` function returns input token as result token. The
next token is instead available in the decode_state, so this change uses
this instead.

* fix(engine): set batch_size and sequence_length

* fix(Jetstream PT): correct warmup internal params

* test(tgi): added a warmup test

* chore(jetstream pt): check input type in decode

* fix(token selector): seed can be a very large number

Before, we could have an error is the seed was bigger than a 64 bit
number.

* fix(Jetstream PT): handle slot's seed in a clean way

* feat(docker): TGI image now include Jetstream Pytorch dependencies

This allows testing TGI images with Jetstream Pytorch.

* fix(Jetstream Pt): batch returned in prefill initialized to None

This is required when there are no more tokens generated after prefill.

* feat(Jetstream Pt): speed-up prefill by avoiding redundant compilation

A new slot is created at each prefill request, and its selector is
passed as argument to a jitted function. The problem is that each new
slot has a new signature, even if the contents are the same. The
solution is to wrap that in a singleton slot object for the prefill, so
the compiler will always see the same object and stop recompiling.

* chore(generator): use prefill bucket sizes defined in Jetstream