Fix ci nightly jetstream (#101)
* ci(nightly): split Pytorch/XLA and Jetstream/Pytorch slow tests

This will simplify debugging in case of failure.

* fix(generator): remove debug log

* fix(ci): workaround to avoid OOM

There seems to be an issue with Jetstream and memory handling. The issue
appears when launching one test after the other, because the previous
memory does not seem to be freed.
A workaround can be splitting tests, that is what this does.