fix(jetstream): correct sampling for jetstream
The Jetstream/Pt engine allows to pass a callback when using the prefill
and generate methods. This callback is used to sample the generated
token with custom function, but the caller function is JIT'ed, making a
strong constraint on the callback signature. So far the callback was
compiled on the first call, making it impossible to change the sampling
algorithm on different requests.
This commit fixes this issue by subclassing the PytorchEngine class and
defining a new `prefill_ex` method that is not JIT'ed. The model calls
are still compiled, so the performance should not be noticeably
affected.