fix(tests): remove expected results from tests with do_sample
Some tests result change when operations are done in a slightly
different way. This has happened now with the torch xla tests, resulting
in different results on the CI.
To avoid this, now tests compare the obtained token and text is
different from the one obtained when running with greedy search.