Jetstream by default (#118)
* test(slots): add unit tests for slots for jetstream too

Implementation is slightly different, so a separate test is added.

* test(truncate): adapt test for jetstream too

* refactor(test): make tinyllama test work for Jetstream and Torch/XLA

Most tests work for both, except for the continuous batching one.
This allows to remove the old GPT2 based tests, that are quite slow and
do not use any sharding or KV cache, so they might not really be
representative of most relevant models on TGI.

* test(gpt2): remove old test

There are equivalent tests now on the TinyLlama model, that run faster,
use the KV cache and sharding.
The only test that does not have an equivalence is the continuous
batching one, but the test was not working for most other models, so I
prefer to remove it anyway, as having it passing was not representative
anyway of the current state.

* feat(tgi): Jetstream/Pytorch is now the default engine

Now that the engine is stable and tested, its engine is set as the
default one for TGI.

* review(test): refactor slot test to avoid repeating code

* feat(tests): use pytests markers to filter jetstream and torch xla tests

So far filtering was done using the name of the test. Now the selection
is done using a custom marker, that allows for clearer filtering.

* review(tests): skip test message clarification

* ci(torch xla): use JETSTREAM_PT_DISABLE env var in command line

For some reason the env var was not carried on (though Jetstream was
disabled anyway). Moving the variable to the command line invocation
will remove a warning in the logs.

* review(ci): fix JETSTREAM_PT_DISABLE env var usage again

* fix(tests): remove expected results from tests with do_sample

Some tests result change when operations are done in a slightly
different way. This has happened now with the torch xla tests, resulting
in different results on the CI.
To avoid this, now tests compare the obtained token and text is
different from the one obtained when running with greedy search.