⬆️ Update dependencies (#120)
* chore(fsdp): remove a warning not relevant anymore

A warning appeared on on an old version of torch xla (2.3.0), but that
is not supported anymore.

* chore: update jetstream dependency to v0.2.4

* chore(docker): increase ulimit to avoid error

When building the docker container, sometimes an error occurs due to a
"too many files open" error. Increasing the ulimit makes the error
disappear.

* fix(docker): "AS" statement should be uppercase to avoid warning

* chore: update TGI dependency to v2.4.1

Also align Dockerfile to TGI's one.

* chore(docker): update accelerate to v1.1.1

* fix(jetstream): correct Gemma and Mixtral config handling

The config object variable for these models was used by the Jetstream
code, but it does not completely match with HF's config definitions.
This creates a class that heritates from both classes, and makes the
adjustments necessary to avoid errors.

* chore(optimum): remove AutoModelForCausalLM from optimum.tpu

It is still possible to import it importing modeling, but it will reduce
the possibility of importing transformers and torch xla before xla2.

* chore: update torch and torch_xla to v2.5.1

* chore(jetstream): token selector operations are done in torch

Conversions of scores tensors from jax to torch and back are done when
calling logits processor. This will be required in newer versions of
transformers.

* chore(dependencies): update transformers to v4.46.3

* chore: update safetensors to v0.4.5

This is to be coherent with accelerate dependencies, and to update to a
newer version.

* review(mixtral): use properties in config to avoid aliasing ambiguity

Instead of assigning separate variables for Jetstream's config class,
properties are added, resulting in accessing the same data and avoiding
ambiguity.