🔂 Fix repetition penalty (#125)
* fix(jetstream): allow to use repetition penalty without crashing

There was an error occurring because the logits processor was expecting
input_ids to be a torch tensor.

* refactor(jetstream): store input tokens in slot as torch.tensors

Since they are mostly used in the logits processor, this will prevent
having to convert them to torch tensor at each decode step.
Functionally it should be the same, but slightly more efficient.

* fix(test): tensor type in test_generator_slot is now Pytorch