🦙 Newer Llamas support (#129)
* feat(jetstream): support RoPE scaling args

* feat(Jetstream): add an explicit error in Llama if MLP bias is enabled

The original Llama model has MLP bias disabled, but it might happen that
it is enabled for finetuned models: in this case an error will appear
when serving with Jetstream Pytorch.

* fix(Jetstream): support tied word embeddings

If the config is enabled, on the Jetstream Pt TGI we create an alias,
that will eventually result in untied weights (but at least it will
work).

* chore(test): add small comment on test to explain why we pick a model

* chore(jetstream): support avoids checking if xla was imported before

This does not seem to be necessary anymore.

* fix(tests): marker names retrieval should be done with iter_markers

Otherwise it only uses the last marker.

* test(llama): add test to show support for Llama 3.2

* test(llama): add quantized test for Llama 3.3-70B

* doc(readme): update to mention we support newer Llama versions with TGI