fix(tgi): set max_batch_prefill_tokens
Starting from TGI 2.4.1, the evaluation of the default value for
max_batch_prefill_tokens in the TGI launcher has changed, leading it
to be set to a default value of 4096 on tpu, while it was previously
set to max_batch_size * max_input_tokens.

This is now fixed in the entrypoint, pending a fix in the launcher.