Bump TGI version to v3.0.0 (#135)
* chore(ci): remove TGI_VERSION argument from workflow

The Dockerfile has a default value, it is easier to only maintain that.

* feat(TGI): update to v3.0.0

Update to TGI 3.0.0, using a simplified Cargo.toml.
This is based on the work done on optimum-neuron:
https://github.com/huggingface/optimum-neuron/pull/748

* fix(tgi): add merge_lora kwarg to download_weights

* fix(tgi): return the correct FinishReason on stop string

* fix(tgi): set max_batch_prefill_tokens

Starting from TGI 2.4.1, the evaluation of the default value for
max_batch_prefill_tokens in the TGI launcher has changed, leading it
to be set to a default value of 4096 on tpu, while it was previously
set to max_batch_size * max_input_tokens.

This is now fixed in the entrypoint, pending a fix in the launcher.

* review(docker): unset MAX_INPUT_LENGTH when set